Deploy a production-ready predictive Linear Mixed Model model

This step allows one to

Load a saved model from LinearMixedModelDevelopment

Run the model against test data to generate predictions

Push these predictions to SQL Server

The linear mixed model functionality works best with smaller data sets.

LinearMixedModelDeployment(type, df, grainCol, personCol,
predictedCol, impute, debug, cores, modelName)

Arguments

type
The type of model (either 'regression' or 'classification')

df
Dataframe whose columns are used for new predictions. Data structure should match development as much as possible. Number of columns, names, types, grain, and predicted must be the same.

grainCol
Optional. The dataframe's column that has IDs pertaining to the grain. No ID columns are truly needed for this step.

personCol
The data frame's columns that represents the patient/person

predictedCol
Column that you want to predict.

impute
For training df, set all-column imputation to T or F. If T, this uses values calculated in development. F leads to removal of rows containing NULLs and is not recommended.

debug
Provides the user extended output to the console, in order to monitor the calculations throughout. Use T or F.

cores
Number of cores you'd like to use. Defaults to 2.

modelName
Optional string. Can specify the model name. If used, you must load the same one in the deploy step.

Format

An object of class R6ClassGenerator of length 24.

Methods

The above describes params for initializing a new linearMixedModelDeployment class with $new(). Individual methods are documented below.

<code>$new()</code>

Initializes a new linear mixed model deployment class using the parameters saved in p, documented above. This method loads, cleans, and prepares data for generating predictions. Usage: $new(p)

<code>$deploy()</code>

Generate new predictions, calculate top factors, and prepare the output dataframe. Usage:$deploy()

<code>$getTopFactors()</code>

Return the grain, all top factors, and their weights. Usage: $getTopFactors(numberOfFactors = NA, includeWeights = FALSE) Params: - numberOfFactors: returns the top n factors. Defaults to all factors. - includeWeights: If TRUE, returns weights associated with each factor.

<code>$getOutDf()</code>

Returns the output dataframe. Usage: $getOutDf()

See also

healthcareai writeData selectData

Examples

#### Classification Example using csv data ####
## 1. Loading data and packages.
ptm <- proc.time()
library(healthcareai)

# setwd('C:/Yourscriptlocation/Useforwardslashes') # Uncomment if using csv

# Can delete this line in your work
csvfile <- system.file("extdata",
                       "HCRDiabetesClinical.csv",
                       package = "healthcareai")

# Replace csvfile with 'path/file'
df <- read.csv(file = csvfile,
               header = TRUE,
               na.strings = c("NULL", "NA", ""))

# Save a dataframe for validation later on
dfDeploy <- df[951:1000,]

## 2. Train and save the model using DEVELOP
print('Historical, development data:')

#> [1] "Historical, development data:"

str(df)

#> 'data.frame':	1000 obs. of  7 variables:
#>  $ PatientEncounterID : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ PatientID          : int  10001 10001 10001 10002 10002 10002 10002 10003 10003 10003 ...
#>  $ SystolicBPNBR      : int  167 153 170 187 188 185 189 149 155 160 ...
#>  $ LDLNBR             : int  195 214 191 135 125 178 101 160 144 130 ...
#>  $ A1CNBR             : num  4.2 5 4 4.4 4.3 5 4 5 6.6 8 ...
#>  $ GenderFLG          : Factor w/ 2 levels "F","M": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ ThirtyDayReadmitFLG: Factor w/ 2 levels "N","Y": 1 1 1 1 1 1 1 1 1 2 ...

set.seed(42)
p <- SupervisedModelDevelopmentParams$new()
p$df <- df
p$type <- "classification"
p$impute <- TRUE
p$grainCol <- "PatientEncounterID"
p$personCol <- "PatientID"
p$predictedCol <- "ThirtyDayReadmitFLG"
p$debug <- FALSE
p$cores <- 1

# Run Linear Mixed Model
lmm <- LinearMixedModelDevelopment$new(p)
lmm$run()

#> Warning: Model is nearly unidentifiable: large eigenvalue ratio
#>  - Rescale variables?

#> Area under the ROC curve is: 0.93 
#> Ideal cutoff is 0.21, yielding TPR of 0.80 and FPR of 0.07 
#> 95% CI AU_ROC: (0.88 , 0.99) 
#> 
#> Area under the PR curve is: 0.81 
#> Ideal cutoff is 0.59, yielding Precision of 0.94 and Recall of 0.68 
#> 

## 3. Load saved model and use DEPLOY to generate predictions. 
print('Fake production data:')

#> [1] "Fake production data:"

str(dfDeploy)

#> 'data.frame':	50 obs. of  7 variables:
#>  $ PatientEncounterID : int  951 952 953 954 955 956 957 958 959 960 ...
#>  $ PatientID          : int  10152 10152 10152 10153 10153 10153 10153 10153 10153 10153 ...
#>  $ SystolicBPNBR      : int  125 112 111 166 161 165 176 163 167 161 ...
#>  $ LDLNBR             : int  122 144 145 119 106 118 145 112 123 106 ...
#>  $ A1CNBR             : num  6.4 6 6 8 8 8 8 8 8 8 ...
#>  $ GenderFLG          : Factor w/ 2 levels "F","M": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ ThirtyDayReadmitFLG: Factor w/ 2 levels "N","Y": 1 1 1 2 2 2 2 2 2 2 ...

p2 <- SupervisedModelDeploymentParams$new()
p2$type <- "classification"
p2$df <- dfDeploy
p2$grainCol <- "PatientEncounterID"
p2$personCol <- "PatientID"
p2$predictedCol <- "ThirtyDayReadmitFLG"
p2$impute <- TRUE
p2$debug <- FALSE
p2$cores <- 1

dL <- LinearMixedModelDeployment$new(p2)

#> Loading Model Info... 
#> The modelName parameter was not specified. Using defaults:
#> - rmodel_probability_LMM.rda
#> - rmodel_info_LMM.rda
#> Loading Data... 

dL$deploy()

#> Model metadata written to predictionMetadata_default_2017-10-24_14.57.10.669.txt
#> WARNING: **This file may contain PHI.**

dfOut <- dL$getOutDf()
head(dfOut)

#>   BindingID BindingNM         LastLoadDTS PatientEncounterID PredictedProbNBR
#> 1         0         R 2017-10-24 14:57:10                951       0.27505247
#> 2         0         R 2017-10-24 14:57:10                952       0.10272940
#> 3         0         R 2017-10-24 14:57:10                953       0.09806673
#> 4         0         R 2017-10-24 14:57:10                954       0.98843802
#> 5         0         R 2017-10-24 14:57:10                955       0.98436055
#> 6         0         R 2017-10-24 14:57:10                956       0.98776848
#>   Factor1TXT    Factor2TXT  Factor3TXT
#> 1     A1CNBR SystolicBPNBR GenderFLG.M
#> 2     A1CNBR SystolicBPNBR GenderFLG.M
#> 3     A1CNBR SystolicBPNBR GenderFLG.M
#> 4     A1CNBR SystolicBPNBR GenderFLG.M
#> 5     A1CNBR SystolicBPNBR GenderFLG.M
#> 6     A1CNBR SystolicBPNBR GenderFLG.M

# Write to CSV (or JSON, MySQL, etc) using plain R syntax
# write.csv(dfOut,'path/predictionsfile.csv')

print(proc.time() - ptm)

#>    user  system elapsed 
#>   3.027   0.079   2.077 

# NOT RUN {
#### Classification example using SQL Server data ####
# This example requires you to first create a table in SQL Server
# If you prefer to not use SAMD, execute this in SSMS to create output table:
# CREATE TABLE dbo.HCRDeployClassificationBASE(
#   BindingID float, BindingNM varchar(255), LastLoadDTS datetime2,
#   PatientEncounterID int, <--change to match inputID
#   PredictedProbNBR decimal(38, 2),
#   Factor1TXT varchar(255), Factor2TXT varchar(255), Factor3TXT varchar(255)
# )

## 1. Loading data and packages.
ptm <- proc.time()
library(healthcareai)

connection.string <- "
driver={SQL Server};
server=localhost;
database=SAM;
trusted_connection=true
"

query <- "
SELECT
[PatientEncounterID]
,[PatientID]
,[SystolicBPNBR]
,[LDLNBR]
,[A1CNBR]
,[GenderFLG]
,[ThirtyDayReadmitFLG]
FROM [SAM].[dbo].[HCRDiabetesClinical]
"

df <- selectData(connection.string, query)

# Save a dataframe for validation later on
dfDeploy <- df[951:1000,]

## 2. Train and save the model using DEVELOP
print('Historical, development data:')
str(df)

set.seed(42)
p <- SupervisedModelDevelopmentParams$new()
p$df <- df
p$type <- "classification"
p$impute <- TRUE
p$grainCol <- "PatientEncounterID"
p$personCol <- "PatientID"
p$predictedCol <- "ThirtyDayReadmitFLG"
p$debug <- FALSE
p$cores <- 1

# Run Linear Mixed Model
lmm <- LinearMixedModelDevelopment$new(p)
lmm$run()

## 3. Load saved model and use DEPLOY to generate predictions. 
print('Fake production data:')
str(dfDeploy)

p2 <- SupervisedModelDeploymentParams$new()
p2$type <- "classification"
p2$df <- dfDeploy
p2$grainCol <- "PatientEncounterID"
p2$personCol <- "PatientID"
p2$predictedCol <- "ThirtyDayReadmitFLG"
p2$impute <- TRUE
p2$debug <- FALSE
p2$cores <- 1

dL <- LinearMixedModelDeployment$new(p2)
dL$deploy()
dfOut <- dL$getOutDf()

writeData(MSSQLConnectionString = connection.string,
          df = dfOut,
          tableName = 'HCRDeployClassificationBASE')

print(proc.time() - ptm)
# }

# NOT RUN {
#### Regression Example using SQL Server data ####
# This example requires you to first create a table in SQL Server
# If you prefer to not use SAMD, execute this in SSMS to create output table:
# CREATE TABLE dbo.HCRDeployRegressionBASE(
#   BindingID float, BindingNM varchar(255), LastLoadDTS datetime2,
#   PatientEncounterID int, <--change to match inputID
#   PredictedValueNBR decimal(38, 2),
#   Factor1TXT varchar(255), Factor2TXT varchar(255), Factor3TXT varchar(255)
# )

## 1. Loading data and packages.
ptm <- proc.time()
library(healthcareai)

connection.string <- "
driver={SQL Server};
server=localhost;
database=SAM;
trusted_connection=true
"

query <- "
SELECT
[PatientEncounterID]
,[PatientID]
,[SystolicBPNBR]
,[LDLNBR]
,[A1CNBR]
,[GenderFLG]
,[ThirtyDayReadmitFLG]
FROM [SAM].[dbo].[HCRDiabetesClinical]
"

df <- selectData(connection.string, query)

# Save a dataframe for validation later on
dfDeploy <- df[951:1000,]

## 2. Train and save the model using DEVELOP
print('Historical, development data:')
str(df)

set.seed(42)
p <- SupervisedModelDevelopmentParams$new()
p$df <- df
p$type <- "regression"
p$impute <- TRUE
p$grainCol <- "PatientEncounterID"
p$personCol <- "PatientID"
p$predictedCol <- "A1CNBR"
p$debug <- FALSE
p$cores <- 1

# Run Linear Mixed Model
lmm <- LinearMixedModelDevelopment$new(p)
lmm$run()

## 3. Load saved model and use DEPLOY to generate predictions. 
dfDeploy$A1CNBR <- NULL # You won't know the response in production
print('Fake production data:')
str(dfDeploy)

p2 <- SupervisedModelDeploymentParams$new()
p2$type <- "regression"
p2$df <- dfDeploy
p2$grainCol <- "PatientEncounterID"
p2$personCol <- "PatientID"
p2$predictedCol <- "A1CNBR"
p2$impute <- TRUE
p2$debug <- FALSE
p2$cores <- 1

dL <- LinearMixedModelDeployment$new(p2)
dL$deploy()
dfOut <- dL$getOutDf()

writeData(MSSQLConnectionString = connection.string,
          df = dfOut,
          tableName = 'HCRDeployRegressionBASE')

print(proc.time() - ptm)
# }

# NOT RUN {
#### Classification example pulling from CSV and writing to SQLite ####

## 1. Loading data and packages.
ptm <- proc.time()
library(healthcareai)

# Can delete these system.file lines in your work
csvfile <- system.file("extdata",
                       "HCRDiabetesClinical.csv",
                       package = "healthcareai")

sqliteFile <- system.file("extdata",
                          "unit-test.sqlite",
                          package = "healthcareai")

# Read in CSV; replace csvfile with 'path/file'
df <- read.csv(file = csvfile,
               header = TRUE,
               na.strings = c("NULL", "NA", ""))

# Save a dataframe for validation later on
dfDeploy <- df[951:1000,]

## 2. Train and save the model using DEVELOP
print('Historical, development data:')
str(df)

set.seed(42)
p <- SupervisedModelDevelopmentParams$new()
p$df <- df
p$type <- "classification"
p$impute <- TRUE
p$grainCol <- "PatientEncounterID"
p$personCol <- "PatientID"
p$predictedCol <- "ThirtyDayReadmitFLG"
p$debug <- FALSE
p$cores <- 1

# Run Linear Mixed Model
lmm <- LinearMixedModelDevelopment$new(p)
lmm$run()

## 3. Load saved model and use DEPLOY to generate predictions. 
print('Fake production data:')
str(dfDeploy)

p2 <- SupervisedModelDeploymentParams$new()
p2$type <- "classification"
p2$df <- dfDeploy
p2$grainCol <- "PatientEncounterID"
p2$personCol <- "PatientID"
p2$predictedCol <- "ThirtyDayReadmitFLG"
p2$impute <- TRUE
p2$debug <- FALSE
p2$cores <- 1

dL <- LinearMixedModelDeployment$new(p2)
dL$deploy()
dfOut <- dL$getOutDf()

writeData(SQLiteFileName = sqliteFile,
          df = dfOut,
          tableName = 'HCRDeployClassificationBASE')

print(proc.time() - ptm)
# }

# NOT RUN {
#### Regression example pulling from CSV and writing to SQLite ####

## 1. Loading data and packages.
ptm <- proc.time()
library(healthcareai)

# Can delete these system.file lines in your work
csvfile <- system.file("extdata",
                       "HCRDiabetesClinical.csv",
                       package = "healthcareai")

sqliteFile <- system.file("extdata",
                          "unit-test.sqlite",
                          package = "healthcareai")

# Read in CSV; replace csvfile with 'path/file'
df <- read.csv(file = csvfile,
               header = TRUE,
               na.strings = c("NULL", "NA", ""))

# Save a dataframe for validation later on
dfDeploy <- df[951:1000,]

## 2. Train and save the model using DEVELOP
print('Historical, development data:')
str(df)

set.seed(42)
p <- SupervisedModelDevelopmentParams$new()
p$df <- df
p$type <- "regression"
p$impute <- TRUE
p$grainCol <- "PatientEncounterID"
p$personCol <- "PatientID"
p$predictedCol <- "A1CNBR"
p$debug <- FALSE
p$cores <- 1

# Run Linear Mixed Model
lmm <- LinearMixedModelDevelopment$new(p)
lmm$run()

## 3. Load saved model and use DEPLOY to generate predictions. 
dfDeploy$A1CNBR <- NULL # You won't know the response in production
print('Fake production data:')
str(dfDeploy)

p2 <- SupervisedModelDeploymentParams$new()
p2$type <- "regression"
p2$df <- dfDeploy
p2$grainCol <- "PatientEncounterID"
p2$personCol <- "PatientID"
p2$predictedCol <- "A1CNBR"
p2$impute <- TRUE
p2$debug <- FALSE
p2$cores <- 1

dL <- LinearMixedModelDeployment$new(p2)
dL$deploy()
dfOut <- dL$getOutDf()

writeData(SQLiteFileName = sqliteFile,
          df = dfOut,
          tableName = 'HCRDeployRegressionBASE')

print(proc.time() - ptm)
# }