Deploy a production-ready predictive XGBoost model

This step allows one to

Automatically load a saved model from XGBoostDevelopment

Run the model against test data to generate predictions

Push these predictions to SQL Server or CSV

XGBoostDeployment(type, df, grainCol,
predictedCol, impute, debug, cores, modelName)

Arguments

type
The type of model (must be multiclass)

df
Dataframe whose columns are used for new predictions. Data structure should match development as much as possible. Number of columns, names, types, grain, and predicted must be the same.

grainCol
The dataframe's column that has IDs pertaining to the grain

predictedCol
Column that you want to predict.

impute
For training df, set all-column imputation to T or F. If T, this uses values calculated in development. F leads to removal of rows containing NULLs and is not recommended.

debug
Provides the user extended output to the console, in order to monitor the calculations throughout. Use T or F.

cores
Number of cores you'd like to use. Defaults to 2.

modelName
Optional string. Can specify the model name. If used, you must load the same one in the deploy step.

Format

An object of class R6ClassGenerator of length 24.

Methods

The above describes params for initializing a new XGBoostDeployment class with $new(). Individual methods are documented below.

<code>$new()</code>

Initializes a new XGBoost deployment class using the parameters saved in p, documented above. This method loads, cleans, and prepares data for generating predictions. Usage: $new(p)

<code>$deploy()</code>

Generate new predictions and prepare the output dataframe. Usage: $deploy()

<code>$getPredictions()</code>

Return the grain and predictions for each class. Usage: $getPredictions()

<code>$getOutDf()</code>

Returns a dataframe containing the grain column, the top 3 probabilities for each row, and the classes associated with those probabilities. Usage: $getOutDf()

See also

healthcareai writeData selectData

Examples

#### Example using csv dataset ####
ptm <- proc.time()
library(healthcareai)

# 1. Load data. Categorical columns should be characters.
# can delete these system.file lines in your work
csvfile <- system.file("extdata",
                      "dermatology_multiclass_data.csv",
                      package = "healthcareai")
# Read in CSV; replace csvfile with 'path/file'
df <- read.csv(file = csvfile,
              header = TRUE,
             stringsAsFactors = FALSE,
              na.strings = c("NULL", "NA", "", "?"))

str(df) # check the types of columns

#> 'data.frame':	366 obs. of  36 variables:
#>  $ PatientID: int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ x1       : chr  "Class2" "Class3" "Class2" "Class2" ...
#>  $ x2       : chr  "Class2" "Class3" "Class1" "Class2" ...
#>  $ x3       : chr  "Class0" "Class3" "Class2" "Class2" ...
#>  $ x4       : chr  "Class3" "Class2" "Class3" "Class0" ...
#>  $ x5       : chr  "Class0" "Class1" "Class1" "Class0" ...
#>  $ x6       : chr  "Class0" "Class0" "Class3" "Class0" ...
#>  $ x7       : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x8       : chr  "Class0" "Class0" "Class3" "Class0" ...
#>  $ x9       : chr  "Class1" "Class1" "Class0" "Class3" ...
#>  $ x10      : chr  "Class0" "Class1" "Class0" "Class2" ...
#>  $ x11      : chr  "Class0" "Class1" "Class0" "Class0" ...
#>  $ x12      : chr  "Class0" "Class0" "Class1" "Class0" ...
#>  $ x13      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x14      : chr  "Class0" "Class1" "Class0" "Class3" ...
#>  $ x15      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x16      : chr  "Class3" "Class1" "Class1" "Class0" ...
#>  $ x17      : chr  "Class2" "Class2" "Class2" "Class2" ...
#>  $ x18      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x19      : chr  "Class0" "Class2" "Class2" "Class3" ...
#>  $ x20      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x21      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x22      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x23      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x24      : chr  "Class0" "Class1" "Class0" "Class0" ...
#>  $ x25      : chr  "Class0" "Class0" "Class2" "Class0" ...
#>  $ x26      : chr  "Class0" "Class0" "Class0" "Class3" ...
#>  $ x27      : chr  "Class0" "Class0" "Class2" "Class0" ...
#>  $ x28      : chr  "Class3" "Class0" "Class3" "Class0" ...
#>  $ x29      : chr  "Class0" "Class0" "Class2" "Class0" ...
#>  $ x30      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x31      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x32      : chr  "Class1" "Class1" "Class2" "Class3" ...
#>  $ x33      : chr  "Class0" "Class0" "Class3" "Class0" ...
#>  $ x34      : int  55 8 26 40 45 41 18 57 22 30 ...
#>  $ target   : chr  "two" "one" "three" "one" ...

dfDeploy <- df[347:366,] # reserve 20 rows for deploy step.

# 2. Develop and save model (saving is automatic)
set.seed(42)
p <- SupervisedModelDevelopmentParams$new()
p$df <- df
p$type <- "multiclass"
p$impute <- TRUE
p$grainCol <- "PatientID"
p$predictedCol <- "target"
p$debug <- FALSE
p$cores <- 1
# xgb_params must be a list with all of these things in it. 
# if you would like to tweak parameters, go for it! 
# Leave objective and eval_metric as they are.
p$xgb_params <- list("objective" = "multi:softprob",
                     "eval_metric" = "mlogloss",
                     "max_depth" = 6, # max depth of each learner
                     "eta" = 0.1, # learning rate
                     "silent" = 0, # verbose output when set to 1
                     "nthread" = 2) # number of processors to use

# Run model
boost <- XGBoostDevelopment$new(p)

#> Warning: Each of the following categorical variable levels occurs 3 times or fewer:
#> -  x6 : Class1
#> -  x27 : Class1
#> -  x33 : Class1
#> Consider grouping these together with other levels.

#> xgb_params for model training are: 
#> List of 7
#>  $ objective  : chr "multi:softprob"
#>  $ eval_metric: chr "mlogloss"
#>  $ max_depth  : num 6
#>  $ eta        : num 0.1
#>  $ silent     : num 0
#>  $ nthread    : num 2
#>  $ num_class  : int 6
#>  

boost$run()

#> Preparing data... 
#> Building model... 
#> Generating predictions... 
#> Generating confusion matrix... 
#> Confusion Matrix and Statistics
#> 
#>          True
#> Predicted one two five six three four
#>     one    22   0    0   0     0    0
#>     two     0  12    0   0     0    1
#>     five    0   0   10   0     1    0
#>     six     0   0    0   4     0    0
#>     three   0   0    0   0    13    0
#>     four    0   0    0   0     0    8
#> 
#> Overall Statistics
#>                                           
#>                Accuracy : 0.9718          
#>                  95% CI : (0.9019, 0.9966)
#>     No Information Rate : 0.3099          
#>     P-Value [Acc > NIR] : < 2.2e-16       
#>                                           
#>                   Kappa : 0.9647          
#>  Mcnemar's Test P-Value : NA              
#> 
#> Statistics by Class:
#> 
#>                      Class: one Class: two Class: five Class: six Class: three
#> Sensitivity              1.0000     1.0000      1.0000    1.00000       0.9286
#> Specificity              1.0000     0.9831      0.9836    1.00000       1.0000
#> Pos Pred Value           1.0000     0.9231      0.9091    1.00000       1.0000
#> Neg Pred Value           1.0000     1.0000      1.0000    1.00000       0.9828
#> Prevalence               0.3099     0.1690      0.1408    0.05634       0.1972
#> Detection Rate           0.3099     0.1690      0.1408    0.05634       0.1831
#> Detection Prevalence     0.3099     0.1831      0.1549    0.05634       0.1831
#> Balanced Accuracy        1.0000     0.9915      0.9918    1.00000       0.9643
#>                      Class: four
#> Sensitivity               0.8889
#> Specificity               1.0000
#> Pos Pred Value            1.0000
#> Neg Pred Value            0.9841
#> Prevalence                0.1268
#> Detection Rate            0.1127
#> Detection Prevalence      0.1127
#> Balanced Accuracy         0.9444

## 3. Load saved model (automatic) and use DEPLOY to generate predictions. 
p2 <- SupervisedModelDeploymentParams$new()
p2$type <- "multiclass"
p2$df <- dfDeploy
p2$grainCol <- "PatientID"
p2$predictedCol <- "target"
p2$impute <- TRUE
p2$debug <- FALSE

# Deploy model to make new predictions
boostD <- XGBoostDeployment$new(p2)

#> Initializing XGBoost Deploy... 
#> Loading Model Info... 
#> The modelName parameter was not specified. Using defaults:
#> - rmodel_probability_XGB.rda
#> - rmodel_info_XGB.rda
#> Loading Data... 

boostD$deploy()

#> Loading XGB Model... 
#> Preparing data... 
#> Generating Predictions... 
#> Ordering top probabilities... 
#> Creating ouput dataframe 

#> Model metadata written to predictionMetadata_default_2017-10-24_14.57.25.460.txt
#> WARNING: **This file may contain PHI.**

# Get output dataframe for csv or SQL
outDf <- boostD$getOutDf()
head(outDf)

#>   BindingID BindingNM         LastLoadDTS PatientID PredictedProb1
#> 1         0         R 2017-10-24 14:57:25       347      0.8988032
#> 2         0         R 2017-10-24 14:57:25       348      0.9559382
#> 3         0         R 2017-10-24 14:57:25       349      0.9601908
#> 4         0         R 2017-10-24 14:57:25       350      0.9517408
#> 5         0         R 2017-10-24 14:57:25       351      0.9855043
#> 6         0         R 2017-10-24 14:57:25       352      0.9847853
#>   PredictedClass1 PredictedProb2 PredictedClass2 PredictedProb3 PredictedClass3
#> 1            four    0.074328810             two    0.007478806             one
#> 2             six    0.012727907             two    0.012387180            four
#> 3             six    0.016711587             two    0.009419603            four
#> 4             six    0.020611083             two    0.010608118            four
#> 5             one    0.003406029            four    0.003207223             two
#> 6             one    0.003690927             two    0.003647161            four

# If you want to write to sqlite:
# sqliteFile <- system.file("extdata",
#                          "unit-test.sqlite",
#                         package = "healthcareai")
# writeData(SQLiteFileName = sqliteFile,
#         df = outDf,
#         tableName = "dermatologyDeployMulticlassBASE")

# Write to CSV (or JSON, MySQL, etc) using plain R syntax
# write.csv(df,'path/predictionsfile.csv')

# Get raw predictions if you want
# rawPredictions <- boostD$getPredictions()

# If you have known labels, check your prediction accuracy like this:
# caret::confusionMatrix(true_label,
#              predicted_label,
#              mode = "everything")

print(proc.time() - ptm)

#>    user  system elapsed 
#>   1.925   0.078   0.490 

#### Example pulling from CSV and writing to SQL server ####
# This example requires you to first create a table in SQL Server
# If you prefer to not use SAMD, execute this in SSMS to create output table:
# CREATE TABLE [dbo].[dermatologyDeployClassificationBASE](
# [BindingID] [int] NULL,[BindingNM] [varchar](255) NULL,
# [LastLoadDTS] [datetime2](7) NULL,
# [PatientID] [decimal](38, 0) NULL,
# [PredictedProb1] [decimal](38, 2) NULL,
# [PredictedClass1] [varchar](255) NULL,
# [PredictedProb2] [decimal](38, 2) NULL,
# [PredictedClass2] [varchar](255) NULL,
# [PredictedProb3] [decimal](38, 2) NULL,
# [PredictedClass3] [varchar](255) NULL)

# 1. Load data. Categorical columns should be characters.
csvfile <- system.file("extdata",
                       "dermatology_multiclass_data.csv",
                       package = "healthcareai")

# Replace csvfile with 'path/file'
df <- read.csv(file = csvfile,
               header = TRUE,
               stringsAsFactors = FALSE,
               na.strings = c("NULL", "NA", "", "?"))

str(df) # check the types of columns

#> 'data.frame':	366 obs. of  36 variables:
#>  $ PatientID: int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ x1       : chr  "Class2" "Class3" "Class2" "Class2" ...
#>  $ x2       : chr  "Class2" "Class3" "Class1" "Class2" ...
#>  $ x3       : chr  "Class0" "Class3" "Class2" "Class2" ...
#>  $ x4       : chr  "Class3" "Class2" "Class3" "Class0" ...
#>  $ x5       : chr  "Class0" "Class1" "Class1" "Class0" ...
#>  $ x6       : chr  "Class0" "Class0" "Class3" "Class0" ...
#>  $ x7       : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x8       : chr  "Class0" "Class0" "Class3" "Class0" ...
#>  $ x9       : chr  "Class1" "Class1" "Class0" "Class3" ...
#>  $ x10      : chr  "Class0" "Class1" "Class0" "Class2" ...
#>  $ x11      : chr  "Class0" "Class1" "Class0" "Class0" ...
#>  $ x12      : chr  "Class0" "Class0" "Class1" "Class0" ...
#>  $ x13      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x14      : chr  "Class0" "Class1" "Class0" "Class3" ...
#>  $ x15      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x16      : chr  "Class3" "Class1" "Class1" "Class0" ...
#>  $ x17      : chr  "Class2" "Class2" "Class2" "Class2" ...
#>  $ x18      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x19      : chr  "Class0" "Class2" "Class2" "Class3" ...
#>  $ x20      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x21      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x22      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x23      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x24      : chr  "Class0" "Class1" "Class0" "Class0" ...
#>  $ x25      : chr  "Class0" "Class0" "Class2" "Class0" ...
#>  $ x26      : chr  "Class0" "Class0" "Class0" "Class3" ...
#>  $ x27      : chr  "Class0" "Class0" "Class2" "Class0" ...
#>  $ x28      : chr  "Class3" "Class0" "Class3" "Class0" ...
#>  $ x29      : chr  "Class0" "Class0" "Class2" "Class0" ...
#>  $ x30      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x31      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x32      : chr  "Class1" "Class1" "Class2" "Class3" ...
#>  $ x33      : chr  "Class0" "Class0" "Class3" "Class0" ...
#>  $ x34      : int  55 8 26 40 45 41 18 57 22 30 ...
#>  $ target   : chr  "two" "one" "three" "one" ...

dfDeploy <- df[347:366,] # reserve 20 rows for deploy step.

# 2. Develop and save model (saving is automatic)
set.seed(42)
p <- SupervisedModelDevelopmentParams$new()
p$df <- df
p$type <- "multiclass"
p$impute <- TRUE
p$grainCol <- "PatientID"
p$predictedCol <- "target"
p$debug <- FALSE
p$cores <- 1
# xgb_params must be a list with all of these things in it. 
# if you would like to tweak parameters, go for it! 
# Leave objective and eval_metric as they are.
p$xgb_params <- list("objective" = "multi:softprob",
                   "eval_metric" = "mlogloss",
                   "max_depth" = 6, # max depth of each learner
                   "eta" = 0.1, # learning rate
                   "silent" = 0, # verbose output when set to 1
                   "nthread" = 2) # number of processors to use

# Run model
boost <- XGBoostDevelopment$new(p)

#> Warning: Each of the following categorical variable levels occurs 3 times or fewer:
#> -  x6 : Class1
#> -  x27 : Class1
#> -  x33 : Class1
#> Consider grouping these together with other levels.

#> xgb_params for model training are: 
#> List of 7
#>  $ objective  : chr "multi:softprob"
#>  $ eval_metric: chr "mlogloss"
#>  $ max_depth  : num 6
#>  $ eta        : num 0.1
#>  $ silent     : num 0
#>  $ nthread    : num 2
#>  $ num_class  : int 6
#>  

boost$run()

#> Preparing data... 
#> Building model... 
#> Generating predictions... 
#> Generating confusion matrix... 
#> Confusion Matrix and Statistics
#> 
#>          True
#> Predicted one two five six three four
#>     one    22   0    0   0     0    0
#>     two     0  12    0   0     0    1
#>     five    0   0   10   0     1    0
#>     six     0   0    0   4     0    0
#>     three   0   0    0   0    13    0
#>     four    0   0    0   0     0    8
#> 
#> Overall Statistics
#>                                           
#>                Accuracy : 0.9718          
#>                  95% CI : (0.9019, 0.9966)
#>     No Information Rate : 0.3099          
#>     P-Value [Acc > NIR] : < 2.2e-16       
#>                                           
#>                   Kappa : 0.9647          
#>  Mcnemar's Test P-Value : NA              
#> 
#> Statistics by Class:
#> 
#>                      Class: one Class: two Class: five Class: six Class: three
#> Sensitivity              1.0000     1.0000      1.0000    1.00000       0.9286
#> Specificity              1.0000     0.9831      0.9836    1.00000       1.0000
#> Pos Pred Value           1.0000     0.9231      0.9091    1.00000       1.0000
#> Neg Pred Value           1.0000     1.0000      1.0000    1.00000       0.9828
#> Prevalence               0.3099     0.1690      0.1408    0.05634       0.1972
#> Detection Rate           0.3099     0.1690      0.1408    0.05634       0.1831
#> Detection Prevalence     0.3099     0.1831      0.1549    0.05634       0.1831
#> Balanced Accuracy        1.0000     0.9915      0.9918    1.00000       0.9643
#>                      Class: four
#> Sensitivity               0.8889
#> Specificity               1.0000
#> Pos Pred Value            1.0000
#> Neg Pred Value            0.9841
#> Prevalence                0.1268
#> Detection Rate            0.1127
#> Detection Prevalence      0.1127
#> Balanced Accuracy         0.9444

## 3. Load saved model (automatic) and use DEPLOY to generate predictions. 
p2 <- SupervisedModelDeploymentParams$new()
p2$type <- "multiclass"
p2$df <- dfDeploy
p2$grainCol <- "PatientID"
p2$predictedCol <- "target"
p2$impute <- TRUE
p2$debug <- FALSE

# Deploy model to make new predictions
boostD <- XGBoostDeployment$new(p2)

#> Initializing XGBoost Deploy... 
#> Loading Model Info... 
#> The modelName parameter was not specified. Using defaults:
#> - rmodel_probability_XGB.rda
#> - rmodel_info_XGB.rda
#> Loading Data... 

boostD$deploy()

#> Loading XGB Model... 
#> Preparing data... 
#> Generating Predictions... 
#> Ordering top probabilities... 
#> Creating ouput dataframe 

#> Model metadata written to predictionMetadata_default_2017-10-24_14.57.26.088.txt
#> WARNING: **This file may contain PHI.**

# Get output dataframe for csv or SQL
outDf <- boostD$getOutDf()
head(outDf)

#>   BindingID BindingNM         LastLoadDTS PatientID PredictedProb1
#> 1         0         R 2017-10-24 14:57:26       347      0.8988032
#> 2         0         R 2017-10-24 14:57:26       348      0.9559382
#> 3         0         R 2017-10-24 14:57:26       349      0.9601908
#> 4         0         R 2017-10-24 14:57:26       350      0.9517408
#> 5         0         R 2017-10-24 14:57:26       351      0.9855043
#> 6         0         R 2017-10-24 14:57:26       352      0.9847853
#>   PredictedClass1 PredictedProb2 PredictedClass2 PredictedProb3 PredictedClass3
#> 1            four    0.074328810             two    0.007478806             one
#> 2             six    0.012727907             two    0.012387180            four
#> 3             six    0.016711587             two    0.009419603            four
#> 4             six    0.020611083             two    0.010608118            four
#> 5             one    0.003406029            four    0.003207223             two
#> 6             one    0.003690927             two    0.003647161            four

# Save the output to SQL server

connection.string <- "
driver={SQL Server};
server=localhost;
database=SAM;
trusted_connection=true
"
writeData(MSSQLConnectionString = connection.string,
       df = outDf,
       tableName = 'dermatologyDeployClassificationBASE')

#> Error: nanodbc/nanodbc.cpp:950: IM002: [unixODBC][Driver Manager]Data source name not found, and no default driver specified 

# Get raw predictions if you want
# rawPredictions <- boostD$getPredictions()
print(proc.time() - ptm)

#>    user  system elapsed 
#>   3.093   0.118   1.121