Create a control chart

Create a control chart, aka Shewhart chart: https://en.wikipedia.org/wiki/Control_chart.

control_chart(d, measure, x, group1, group2, center_line = mean, sigmas = 3,
  save_to, plot_width = 8, plot_height = 4, plot_title = NULL,
  plot_catpion = NULL, plot_font_size = 11)

Arguments

d
data frame or a path to a csv file that will be read in

measure
variable of interest mapped to y-axis (quoted, ie as a string)

x
variable to go on the x-axis, often a time variable. If unspecified row indices will be used (quoted)

group1
Optional grouping variable to be panelled horizontally (quoted)

group2
Optional grouping variable to be panelled vertically (quoted)

center_line
Function used to calculate central tendency. Defaults to mean

sigmas
Number of standard deviations above and below the central tendency to call a point influenced by "special cause variation." Defaults to 3

save_to
Optional file path to save chart. If not provided, the chart will be printed on screen

plot_width
In inches. Only used if save_to is specified

plot_height
In inches. Only used if save_to is specified

plot_title
Title in upper-left

plot_catpion
Caption in lower-right

plot_font_size
Base font size; text elements will be scaled to this

Value

Generally called for the side effect of printing the control chart or writing the control chart to file, depending on whether save_to is specified. Invisibly, returns a ggplot object for further customization.

Examples

# NOT RUN {
# Create a data frame to plot
d <-
  tibble::data_frame(
    day = sample(c("Monday", "Tuesday", "Wednesday", "Thursday", "Friday"),
                 100, TRUE),
    person = sample(c("Tom", "Jane", "Alex"), 100, TRUE),
    count = rbinom(100, 20, ifelse(day == "Friday", .5, .2)),
    date = Sys.Date() - sample.int(100))

# Minimal arguments are the data and the column to put on the y-axis.
# If x is not provided, observations will be plotted in order of the rows

control_chart(d, "count")

# Specify categorical variables for group1 and/or group2 to get a separate
# panel for each category

control_chart(d, "count", group1 = "day", group2 = "person")

# In addition to printing or writing the plot to file, control_chart
# returns the plot as a ggplot2 obejct, which you can then further customize

library(ggplot2)
my_chart <- control_chart(d, "count", "date")
my_chart +
  ylab("Number of Adverse Events") +
  scale_x_date(name = "Week of", date_breaks = "week") +
  theme(axis.text.x = element_text(angle = -90, vjust = 0.5, hjust=1))
# }