Get variable importances

Get variable importances

get_variable_importance(models)

Arguments

models
model_list object

Value

Data frame of variables and their importance for predictive power

Details

Some algorithms provide variable importance, others don't. The best-performing model that offers variable importance will be used.

See also

Examples

m <- flash_models(mtcars, outcome = mpg, models = "rf")

#> mpg looks numeric, so training regression algorithms.

#> Training at fixed values: Random Forest

#> 
#> *** Models successfully trained. The model object contains the training data minus ignored ID columns. ***
#> *** If there was PHI in training data, normal PHI protocols apply to the model object. ***

get_variable_importance(m)

#> # A tibble: 10 x 2
#>    variable importance
#>  * <chr>         <dbl>
#>  1 cyl          100.  
#>  2 disp          73.8 
#>  3 wt            53.7 
#>  4 hp            40.7 
#>  5 vs            34.4 
#>  6 am            30.3 
#>  7 carb          11.5 
#>  8 drat           9.73
#>  9 gear           6.35
#> 10 qsec           0.