Function to suppress specific warnings in unit tests

This function allows one to suppress one or more specific warnings. The intended purpose is to suppress warning messages that are expected when running specific unit tests, but are unrelated to the functionality being tested.

ignoreSpecWarn(code, wRegexps)

Arguments

code
A piece of code to run, for which the warnings should be suppressed

wRegexps
A vector of regular expressions corresponding to the warnings that should be suppressed

References

http://healthcareai-r.readthedocs.io

See also

healthcareai