Machine Learn

Prepare data and train machine learning models.

machine_learn(d, ..., outcome, models = c("rf", "knn"), impute = TRUE)

Arguments

d
A data frame

...
Columns to be ignored in model training, e.g. ID columns, unquoted.

outcome
Name of the target column, i.e. what you want to predict. Unquoted. Must be named, i.e. you must specify outcome =

models
Models to be trained, k-nearest neighbors and random forest by default. See supported_models for details.

impute
Logical, if TRUE (default) missing values will be filled by hcai_impute

Value

model_list object ready to make predictions via predict.model_list

Details

This is a high-level wrapper function. For finer control of data cleaning and preparation use prep_data or the functions it wraps. For finer control of model tuning use tune_models.

Examples

# Split data into training and test sets using a subset of the data for speed
test_data <- pima_diabetes[1:10, ]
training_data <- pima_diabetes[11:100, ]

### Classification ###

# Clean and prep the data, tune algorithms over hyperparameter values to predict diabetes
models <- machine_learn(training_data, outcome = diabetes)

#> Training new data prep recipe

#> diabetes looks categorical, so training classification algorithms.

#> Running cross validation for Random Forest

#> Running cross validation for k-Nearest Neighbors

# Make predictions (predicted probability of diabetes) on test data
predict(models, test_data)

#> Prepping data based on provided recipe

#> healthcareai-predicted data. Column "predicted_diabetes" predicted by Random Forest tuned on ROC. Performance in training: ROC = 0.8.

#> # A tibble: 10 x 11
#>    diabetes predicted_diabe… patient_id pregnancies plasma_glucose diastolic_bp
#>  * <chr>               <dbl>      <int>       <int>          <int>        <int>
#>  1 Y                   0.675          1           6            148           72
#>  2 N                   0.403          2           1             85           66
#>  3 Y                   0.517          3           8            183           64
#>  4 N                   0.208          4           1             89           66
#>  5 Y                   0.492          5           0            137           40
#>  6 N                   0.478          6           5            116           74
#>  7 Y                   0.304          7           3             78           50
#>  8 N                   0.622          8          10            115           NA
#>  9 Y                   0.677          9           2            197           70
#> 10 Y                   0.521         10           8            125           96
#> # ... with 5 more variables: skinfold <int>, insulin <int>, weight_class <chr>,
#> #   pedigree <dbl>, age <int>

### Regression ###

# Predict numeric outcomes simply by specifying the name of the outcome variable
age_model <- machine_learn(training_data, outcome = age)

#> Training new data prep recipe

#> age looks numeric, so training regression algorithms.

#> Running cross validation for Random Forest

#> Running cross validation for k-Nearest Neighbors

# If new data isn't specifed, get predictions on training data
predict(age_model)

#> healthcareai-predicted data. Column "predicted_age" predicted by Random Forest tuned on RMSE. Performance in training: RMSE = 8.82.

#> # A tibble: 90 x 17
#>      age predicted_age patient_id pregnancies plasma_glucose diastolic_bp
#>  * <int>         <dbl>      <int>       <int>          <dbl>        <dbl>
#>  1    30          35.2         11           4            110         92.0
#>  2    34          40.1         12          10            168         74.0
#>  3    57          49.6         13          10            139         80.0
#>  4    59          47.0         14           1            189         60.0
#>  5    51          47.5         15           5            166         72.0
#>  6    32          35.2         16           7            100         73.0
#>  7    31          32.3         17           0            118         84.0
#>  8    31          34.1         18           7            107         74.0
#>  9    33          29.6         19           1            103         30.0
#> 10    32          31.0         20           1            115         70.0
#> # ... with 80 more rows, and 11 more variables: skinfold <dbl>, insulin <dbl>,
#> #   pedigree <dbl>, weight_class_normal <dbl>, weight_class_obese <dbl>,
#> #   weight_class_overweight <dbl>, weight_class_hcai_missing <dbl>,
#> #   weight_class_other <dbl>, diabetes_Y <dbl>, diabetes_other <dbl>,
#> #   diabetes_hcai_missing <dbl>