Patient diabetes dataset

A dataset containing diabetes status and other health-related variables for 768 females, at least 21 years old, of Pima Indian heritage. As pointed out (see source URL below), the source data had some biologically impossible zero values. We have replaced zero values in every variable except Pregnancies with NA.

pima_diabetes

Format

A tibble data frame with 768 rows and 10 variables:

patient_id

unique identifier

pregnancies

Number of times pregnant

plasma_glucose

Plasma glucose concentration 2 hours in an oral glucose tolerance test

diastolic_bp

Diastolic blood pressure (mm Hg)

skinfold

Triceps skin fold thickness (mm)

insulin

2-Hour serum insulin (mu U/ml)

weight_class

Derived from BMI

pedigree

Diabetes pedigree function

age

Age (years)

diabetes

Y/N diagnosis per WHO criteria

Source

https://archive.ics.uci.edu/ml/datasets/pima+indians+diabetes