Patient medications dataset

This is a companion dataset for pima_diabetes. The pima_diabetes dataset is real; this dataset is synthetic. You can see how it was generated here: https://docs.healthcare.ai/articles/site_only/best_levels.html#appendix-data-generation. Briefly, each patient in pima_diabetes is assigned 0-4 medications from the following six: insulin and metformin are more common among diabetics, prednisone and metoprolol are less common among diabetics, and nexium and tiotropium are equally likely among diabetic and non-diabetic patients. Each patient-medication has a years_taken value associated with it, which is a random number drawn from an exponential distribution.

pima_meds

Format

A tibble data frame with 1,604 rows and 3 variables:

patient_id

Unique identifier, used to join pima_diabetes

medication

One of the six medications described above

years_taken

Numeric value indicating the duration the medication has been used

See also