Plot predictions from a classification model

Plot predictions from a classification model

plot_classification_predictions(x, title = NULL,
  fill_colors = c("firebrick", "steelblue"), fill_alpha = 0.7,
  curve_flex = 1, font_size = 11, target)

Arguments

x
hcai_predicted_df from plot.hcai_predicted_df

title
Character: Plot title, default NULL produces no title.

fill_colors
Length-2 character vector: colors to fill density curves. Default is c("firebrick", "steelblue"). If named, names must match unique(x[[target]]), in any order.

fill_alpha
Number in [0, 1] giving opacity of fill colors.

curve_flex
Numeric. Kernal adjustment for density curves. Default is 1. Less than 1 makes curves more flexible, analogous to smaller bins in a histogram; greater than 1 makes curves more rigid.

font_size
Number: Relative size of all font in plot, default = 11

target
outcome column name

Value

ggplot object

Details

This function is not meant to be called directly. Use plot(predictions) instead. You can pass arguments to this function (e.g. a plot title) through plot.