Pull data into R via an ODBC connection

Select data from an ODBC database and return the results as a data frame.

selectData(MSSQLConnectionString = NULL, query, SQLiteFileName = NULL,
  randomize = FALSE)

Arguments

MSSQLConnectionString
A string specifying the driver, server, database, and whether Windows Authentication will be used.

query
The SQL query (in ticks or quotes)

SQLiteFileName
A string. If dbtype is SQLite, here one specifies the database file to query from

randomize
Boolean that dictates whether returned rows are randomized

Value

df A data frame containing the selected rows

References

http://healthcareai-r.readthedocs.io

See also

healthcareai writeData

Examples

# This example is specific to SQL Server

# To instead pull data from Oracle see here 
# https://cran.r-project.org/web/packages/ROracle/ROracle.pdf
# To pull data from MySQL see here 
# https://cran.r-project.org/web/packages/RMySQL/RMySQL.pdf
# To pull data from Postgres see here 
# https://cran.r-project.org/web/packages/RPostgreSQL/RPostgreSQL.pdf

connectionString <- '
  driver={SQL Server};
  server=localhost;
  database=SAM;
  trustedConnection=true
  '

query <- '
  SELECT
    A1CNBR
  FROM SAM.dbo.HCRDiabetesClinical
  '

df <- selectData(connectionString, query)

#> Error: nanodbc/nanodbc.cpp:950: IM002: [unixODBC][Driver Manager]Data source name not found, and no default driver specified 

head(df)

#>                                               
#> 1 function (x, df1, df2, ncp, log = FALSE)    
#> 2 {                                           
#> 3     if (missing(ncp))                       
#> 4         .Call(C_df, x, df1, df2, log)       
#> 5     else .Call(C_dnf, x, df1, df2, ncp, log)
#> 6 }