Sets console logging to a file in the working directory.

Logs all console messages and output to a timestamped file called, "deployConsoleLog..." in the working directory. This is meant to be used in production to monitor console output on ETL servers.

start_prod_logs()

See also

Examples

# NOT RUN {
# Object Attribute Base should contain a script like this:
# Move to working directory
setwd("Prod Server/Machine Learning/MyProject")

# Start console logging
start_prod_logs()

# Deploy -------------------------------------------------
# Test deployment. Comment/delete when working
catalyst_test_deploy_in_prod()

# Deploy model
source("myDeployScript.R")

# Stop console logging
stop_prod_logs()
# }