Date and Time Feature Generator

`step_date_hcai` creates a *specification* of a recipe step that will convert date data into factor or numeric variable(s). This step will guess the date format of columns with the "_DTS" suffix, and then create either `categories` or `continuous` columns. Various portions of this step are copied from `recipes::step_date`.

step_date_hcai(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  feature_type = "continuous",
  columns = NULL,
  skip = FALSE,
  id = rand_id("bagimpute")
)

# S3 method for step_date_hcai
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables that will be used to create the new variables. The selected variables should have class `Date` or `POSIXct` or their name must end with `DTS`. See [selections()] for more details. For the `tidy` method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned? By default, the function assumes that the new variable columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the number of NA values have been counted in preprocessing.

feature_type
character, either `continuous` (default) or `categories`.

columns
A character string of variables that will be used as inputs. This field is a placeholder and will be populated once [prep.recipe()] is used.

skip
A logical. Should the step be skipped when the recipe is baked?

id
a unique step id that will be used to unprep

x
A `step_date_hcai` object.

Value

For `step_date_hcai`, an updated version of recipe with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` (the selectors or variables selected), `value` (the feature names), and `ordinal` (a logical).

Details

Unlike other steps, `step_date_hcai` does *not* remove the original date variables. [step_rm()] can be used for this purpose.

Examples

library(lubridate)

#> 
#> Attaching package: ‘lubridate’

#> The following objects are masked from ‘package:base’:
#> 
#>     date, intersect, setdiff, union

library(recipes)

examples <- data.frame(Dan = ymd("2002-03-04") + days(1:10),
                       Stefan = ymd("2006-01-13") + days(1:10))
date_rec <- recipe(~ Dan + Stefan, examples) %>%
  step_date_hcai(all_predictors())

date_rec <- prep(date_rec, training = examples)

date_values <- bake(date_rec, new_data = examples)
date_values

#> # A tibble: 10 x 12
#>    Dan        Stefan     Dan_dow_sin Dan_dow_cos Dan_month_sin Dan_month_cos
#>    <date>     <date>           <dbl>       <dbl>         <dbl>         <dbl>
#>  1 2002-03-05 2006-01-14    4.34e- 1      -0.901             1      6.12e-17
#>  2 2002-03-06 2006-01-15   -4.34e- 1      -0.901             1      6.12e-17
#>  3 2002-03-07 2006-01-16   -9.75e- 1      -0.223             1      6.12e-17
#>  4 2002-03-08 2006-01-17   -7.82e- 1       0.623             1      6.12e-17
#>  5 2002-03-09 2006-01-18   -2.45e-16       1                 1      6.12e-17
#>  6 2002-03-10 2006-01-19    7.82e- 1       0.623             1      6.12e-17
#>  7 2002-03-11 2006-01-20    9.75e- 1      -0.223             1      6.12e-17
#>  8 2002-03-12 2006-01-21    4.34e- 1      -0.901             1      6.12e-17
#>  9 2002-03-13 2006-01-22   -4.34e- 1      -0.901             1      6.12e-17
#> 10 2002-03-14 2006-01-23   -9.75e- 1      -0.223             1      6.12e-17
#> # … with 6 more variables: Dan_year <dbl>, Stefan_dow_sin <dbl>,
#> #   Stefan_dow_cos <dbl>, Stefan_month_sin <dbl>, Stefan_month_cos <dbl>,
#> #   Stefan_year <dbl>

# changing `feature_type` to `categories`
date_rec <-
  recipe(~ Dan + Stefan, examples) %>%
  step_date_hcai(all_predictors(), feature_type = "categories")

date_rec <- prep(date_rec, training = examples)

date_values <- bake(date_rec, new_data = examples)
date_values

#> # A tibble: 10 x 8
#>    Dan        Stefan     Dan_dow Dan_month Dan_year Stefan_dow Stefan_month
#>    <date>     <date>     <fct>   <fct>        <dbl> <fct>      <fct>       
#>  1 2002-03-05 2006-01-14 Tue     Mar           2002 Sat        Jan         
#>  2 2002-03-06 2006-01-15 Wed     Mar           2002 Sun        Jan         
#>  3 2002-03-07 2006-01-16 Thu     Mar           2002 Mon        Jan         
#>  4 2002-03-08 2006-01-17 Fri     Mar           2002 Tue        Jan         
#>  5 2002-03-09 2006-01-18 Sat     Mar           2002 Wed        Jan         
#>  6 2002-03-10 2006-01-19 Sun     Mar           2002 Thu        Jan         
#>  7 2002-03-11 2006-01-20 Mon     Mar           2002 Fri        Jan         
#>  8 2002-03-12 2006-01-21 Tue     Mar           2002 Sat        Jan         
#>  9 2002-03-13 2006-01-22 Wed     Mar           2002 Sun        Jan         
#> 10 2002-03-14 2006-01-23 Thu     Mar           2002 Mon        Jan         
#> # … with 1 more variable: Stefan_year <dbl>