Last Observation Carried Forward Imputation

step_locfimpute creates a *specification* of a recipe step that will substitute missing values with the most recent variable value. If the first variable value is missing, it is imputed with the first present value.

step_locfimpute(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  skip = FALSE,
  id = rand_id("bagimpute")
)

# S3 method for step_locfimpute
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables will be imputed. See [selections()] for more details. For the `tidy` method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the number of NA values have been counted in preprocessing.

skip
A logical. Should the step be skipped when the recipe is baked?

id
a unique step id that will be used to unprep

x
A `step_locfimpute` object.

Value

For step_locfimpute, an updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected) and trained (a logical that states whether the recipe has been prepped).

Examples

library(recipes)

prepped <-
  recipe(formula = "~.", pima_diabetes) %>%
  step_locfimpute(weight_class, insulin, skinfold, diastolic_bp) %>%
  prep()

bake(prepped, new_data = pima_diabetes)

#> # A tibble: 768 x 10
#>    patient_id pregnancies plasma_glucose diastolic_bp skinfold insulin
#>         <int>       <int>          <int>        <int>    <int>   <int>
#>  1          1           6            148           72       35      94
#>  2          2           1             85           66       29      94
#>  3          3           8            183           64       29      94
#>  4          4           1             89           66       23      94
#>  5          5           0            137           40       35     168
#>  6          6           5            116           74       35     168
#>  7          7           3             78           50       32      88
#>  8          8          10            115           50       32      88
#>  9          9           2            197           70       45     543
#> 10         10           8            125           96       45     543
#> # … with 758 more rows, and 4 more variables: weight_class <fct>,
#> #   pedigree <dbl>, age <int>, diabetes <fct>