Find variation across groups

Compare variation among groups on a continuous measure. Plot boxplot and perform ANOVA with Tukey's HSD statistical test to compare variation in a numeric variable (measureColumn) across groups (categoricalCols).

variationAcrossGroups(df, categoricalCols, measureColumn,
  plotGroupDifferences = FALSE, returnGroupStats = FALSE, dateCol = NULL,
  levelOfDateGroup = "monthly", sigLevel = 0.05)

Arguments

df
A data frame containing group and measure columns.

categoricalCols
Character. Vector containing the name(s) of column(s) to group by.

measureColumn
Character. The name of the numeric variable of interest.

plotGroupDifferences
Optional. Logical. Plot results of Tukey's HSD test: mean differences between groups and confidence intervals for each pairwise group comparison? Default is FALSE.

returnGroupStats
Optional. Logical. In addition to the model summary table, return summary statistics for each group? Default is FALSE

dateCol
Optional. Date. A date(time) column to group by (grouped by month by default).

levelOfDateGroup
Optional. Character. Level at which to group dateCol. One of "yearly", "quarterly", "monthly" (default), or "weekly".

sigLevel
Optional. Numeric value between zero and one giving the alpha value for Tukey HSD test, i.e. the p-value threshold for significance.

Value

By default, a data frame giving summary statistics from Tukey's HSD test. If returnGroupStats is TRUE, a list of two data frames giving model and group summary statistics respectively.

This function always induces the side effect of printing a boxplot to compare variation across groups. If plotGroupDifferences is TRUE, also plots a mean differences and confidence intervals between groups.

References

http://healthcare.ai

This function uses multcompView::multcompLetters() https://CRAN.R-project.org/package=multcompView

See also

healthcareai findVariation

Examples

### Example 1 ###

# This example dataset has two columns: a blood test result (value)
# and an anesthetic treatment (anesthetic). There are five anesthetics in the 
# dataset: Midazolam, Propofol, Ketamine, Thiamylal, and Diazepam.

set.seed(35)
df1 <- data.frame(
  anesthetic = c(rep("Midazolam", 50), rep("Propofol", 20), rep("Ketamine", 40),
                 rep("Thiamylal", 80),  rep("Diazepam", 20)),
  value = c(sample(2:5, 50, replace = TRUE), sample(6:10, 20, replace = TRUE),
            sample(1:7, 40, replace = TRUE), sample(3:10, 80, replace = TRUE),
            sample(10:20, 20, replace = TRUE)))
head(df1)

#>   anesthetic value
#> 1  Midazolam     5
#> 2  Midazolam     2
#> 3  Midazolam     4
#> 4  Midazolam     5
#> 5  Midazolam     3
#> 6  Midazolam     2

variationAcrossGroups(df1, "anesthetic", "value", sigLevel = .01)

#>                 Groups Mean Difference Adjusted p-value
#> 1   Propofol-Midazolam           4.640            0.000
#> 2   Diazepam-Midazolam          12.090            0.000
#> 3    Diazepam-Ketamine          11.875            0.000
#> 4   Diazepam-Thiamylal           9.438            0.000
#> 5    Diazepam-Propofol           7.450            0.000
#> 6    Propofol-Ketamine           4.425            0.000
#> 7  Thiamylal-Midazolam           2.652            0.000
#> 8   Thiamylal-Ketamine           2.437            0.000
#> 9   Propofol-Thiamylal           1.988            0.001
#> 10  Ketamine-Midazolam           0.215            0.986

# The boxplot tells us that Diazepam, Propofol, and Thiamylal all have 
# significantly different mean values from all other groups, including each other
# (p <= 0.01). Midazolam and Ketamine do not have significantly different mean 
# values because they share the label "a", but they are significantly different 
# from all the other treatments.

### Example 2 ###

# This example dataset has three columns: department, admission date, and
# blood pressure reading (BP). We will examine whether blood pressures
# vary by department and year of admission.

set.seed(2017)
n <- 200
bp <- data.frame(department = sample(c("Cardiology", "Oncology", "Gastroenterology"),
                                     size = n,
                                     replace = TRUE,
                                     prob =  c(0.5, 0.3, 0.2)),
                 admit_date = sample(seq(as.Date("2015-01-01"),
                                         as.Date("2017-12-31"),
                                         by = "day"),
                                     size = n))
bp$BP <- floor(rnorm(n,
                     120 *
                       ifelse(bp$admit_date > "2015-12-31", 1.5, 1) +
                       ifelse(bp$department == "Cardiology", 80, 0),
                     ifelse(bp$department == "Oncology", 60, 30)))
head(bp)

#>         department admit_date  BP
#> 1 Gastroenterology 2016-07-09 194
#> 2         Oncology 2017-01-11 221
#> 3       Cardiology 2017-09-28 253
#> 4       Cardiology 2017-07-05 263
#> 5         Oncology 2016-05-31 145
#> 6         Oncology 2017-01-21 199

variationAcrossGroups(bp,
                      categoricalCols = "department",
                      measureColumn = "BP",
                      dateCol = "admit_date",
                      levelOfDateGroup = "yearly",
                      plotGroupDifferences = TRUE)

#>                                         Groups Mean Difference Adjusted p-value
#> 1                Cardiology.2016-Oncology.2015         162.185            0.000
#> 2                Cardiology.2017-Oncology.2015         153.694            0.000
#> 3        Cardiology.2016-Gastroenterology.2015         153.548            0.000
#> 4                Cardiology.2016-Oncology.2017         111.917            0.000
#> 5                Cardiology.2015-Oncology.2015          97.686            0.000
#> 6        Cardiology.2017-Gastroenterology.2015         145.057            0.000
#> 7                Cardiology.2017-Oncology.2017         103.426            0.000
#> 8                  Oncology.2016-Oncology.2015          90.470            0.000
#> 9              Cardiology.2016-Cardiology.2015          64.498            0.000
#> 10         Gastroenterology.2017-Oncology.2015          87.208            0.000
#> 11               Cardiology.2016-Oncology.2016          71.715            0.000
#> 12             Cardiology.2017-Cardiology.2015          56.007            0.000
#> 13       Cardiology.2015-Gastroenterology.2015          89.050            0.000
#> 14         Gastroenterology.2016-Oncology.2015          84.720            0.000
#> 15       Cardiology.2016-Gastroenterology.2017          74.977            0.000
#> 16       Cardiology.2016-Gastroenterology.2016          77.465            0.000
#> 17               Cardiology.2017-Oncology.2016          63.224            0.000
#> 18       Cardiology.2017-Gastroenterology.2017          66.486            0.000
#> 19       Cardiology.2017-Gastroenterology.2016          68.974            0.000
#> 20         Oncology.2016-Gastroenterology.2015          81.833            0.000
#> 21 Gastroenterology.2017-Gastroenterology.2015          78.571            0.000
#> 22 Gastroenterology.2016-Gastroenterology.2015          76.083            0.001
#> 23               Cardiology.2015-Oncology.2017          47.418            0.001
#> 24                 Oncology.2017-Oncology.2015          50.268            0.004
#> 25                 Oncology.2016-Oncology.2017          40.202            0.074
#> 26         Gastroenterology.2017-Oncology.2017          36.940            0.206
#> 27         Oncology.2017-Gastroenterology.2015          41.632            0.228
#> 28         Gastroenterology.2016-Oncology.2017          34.452            0.353
#> 29       Cardiology.2015-Gastroenterology.2016          12.967            0.988
#> 30             Cardiology.2016-Cardiology.2017           8.491            0.995
#> 31       Cardiology.2015-Gastroenterology.2017          10.479            0.996
#> 32               Cardiology.2015-Oncology.2016           7.217            0.999
#> 33         Gastroenterology.2015-Oncology.2015           8.636            1.000
#> 34         Oncology.2016-Gastroenterology.2016           5.750            1.000
#> 35         Oncology.2016-Gastroenterology.2017           3.262            1.000
#> 36 Gastroenterology.2017-Gastroenterology.2016           2.488            1.000

                       
# Since plotGroupDifferences = TRUE and the default of returnGroupStats is
# FALSE, the function prints the boxplot and the 95% family-wise confidence
# interval plot, and returns the summary statistics data frame. The two plots
# show: 
# 
#   1. The boxplot of BP across all combinations of the two categories. 
#   department has 3 levels, as does date grouped by year, so there are a total
#   of 3 x 3 = 9 groupings, which are shown on the x axis of the boxplot. Groups
#   that have a shared letter are *not* significantly different. For example,
#   (Gastroenterology | 2015) and (Oncology | 2015) share a "b" label, so
#   patients in those groups do not have significantly different mean BP. On the
#   other hand, (Cardiology | 2015) and (Gastroenterology | 2015) do not share a
#   label, so patients in those groups do have significantly different BP. Likewise,
#   Oncology patients in 2015 have different BPs from Oncology patients in either of the subsequent
#   years, but Oncology patients in 2016 and 2017 do not have significantly
#   different BP (as shown by their shared "c" label). 
#   
#   2. Confidence-interval level plot This plot present the results
#   of the Tukey's Honest Significant Differences test. It compares all possible
#   pairs of groups and adjusts p-values for multiple comparisons. Red lines
#   indicate a significant difference between the two groups at the chosen 
#   significance level (0.05 by default). Groups are ordered by p-values. The
#   group with the greater mean value is always listed first (e.g. Cardiology |
#   2016 has greater BP than Oncology 2015).