Compare predictive models, created on your data

This step allows you to create an XGBoost classification model, based on your data. Use model type 'multiclass' with 2 or more classes. XGBoost is an ensemble model, well suited to non-linear data and very fast. Can be parameter-dependent.

XGBoostDevelopment(type, df, grainCol, predictedCol,
impute, debug, cores, modelName, xgb_params, xgb_nrounds)

Arguments

type
The type of model. Currently requires 'multiclass'.

df
Dataframe whose columns are used for calc.

grainCol
Optional. The dataframe's column that has IDs pertaining to the grain. No ID columns are truly needed for this step.

predictedCol
Column that you want to predict. If you're doing classification then this should be Y/N.

impute
Set all-column imputation to T or F. If T, this uses mean replacement for numeric columns and most frequent for factorized columns. F leads to removal of rows containing NULLs. Values are saved for deployment.

debug
Provides the user extended output to the console, in order to monitor the calculations throughout. Use T or F.

cores
Number of cores you'd like to use. Defaults to 2.

modelName
Optional string. Can specify the model name. If used, you must load the same one in the deploy step.

xgb_params
A list, containing optional xgboost parameters. The full list of params can be found at http://xgboost.readthedocs.io/en/latest/parameter.html.

xgb_nrounds
Number of rounds to use for boosting.

Format

An object of class R6ClassGenerator of length 24.

Methods

The above describes params for initializing a new XGBoostDevelopment class with $new(). Individual methods are documented below.

<code>$new()</code>

Initializes a new XGBoost development class using the parameters saved in p, documented above. This method loads, cleans, and prepares data for model training. Usage: $new(p)

<code>$run()</code>

Trains model, displays predictions and class-wise performance. Usage: $new()

<code>$getPredictions()</code>

Returns the predictions from test data. Usage: $getPredictions()

<code>$generateConfusionMatrix()</code>

Returns the confusion matrix and statistics generated during model development. Usage: $getConfusionMatrix()

References

http://healthcareai-r.readthedocs.io

See also

Information on the example dataset can be found at: http://archive.ics.uci.edu/ml/datasets/dermatology/

Information on the xgboost parameters can be found at: https://github.com/dmlc/xgboost/blob/master/doc/parameter.md selectData

Examples

#### Example using csv dataset ####
ptm <- proc.time()
library(healthcareai)

# 1. Load data. Categorical columns should be characters.
csvfile <- system.file("extdata",
                      "dermatology_multiclass_data.csv",
                      package = "healthcareai")

 # Replace csvfile with 'path/file'
df <- read.csv(file = csvfile,
              header = TRUE,
              stringsAsFactors = FALSE,
              na.strings = c("NULL", "NA", "", "?"))

str(df) # check the types of columns

#> 'data.frame':	366 obs. of  36 variables:
#>  $ PatientID: int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ x1       : chr  "Class2" "Class3" "Class2" "Class2" ...
#>  $ x2       : chr  "Class2" "Class3" "Class1" "Class2" ...
#>  $ x3       : chr  "Class0" "Class3" "Class2" "Class2" ...
#>  $ x4       : chr  "Class3" "Class2" "Class3" "Class0" ...
#>  $ x5       : chr  "Class0" "Class1" "Class1" "Class0" ...
#>  $ x6       : chr  "Class0" "Class0" "Class3" "Class0" ...
#>  $ x7       : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x8       : chr  "Class0" "Class0" "Class3" "Class0" ...
#>  $ x9       : chr  "Class1" "Class1" "Class0" "Class3" ...
#>  $ x10      : chr  "Class0" "Class1" "Class0" "Class2" ...
#>  $ x11      : chr  "Class0" "Class1" "Class0" "Class0" ...
#>  $ x12      : chr  "Class0" "Class0" "Class1" "Class0" ...
#>  $ x13      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x14      : chr  "Class0" "Class1" "Class0" "Class3" ...
#>  $ x15      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x16      : chr  "Class3" "Class1" "Class1" "Class0" ...
#>  $ x17      : chr  "Class2" "Class2" "Class2" "Class2" ...
#>  $ x18      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x19      : chr  "Class0" "Class2" "Class2" "Class3" ...
#>  $ x20      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x21      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x22      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x23      : chr  "Class0" "Class2" "Class0" "Class2" ...
#>  $ x24      : chr  "Class0" "Class1" "Class0" "Class0" ...
#>  $ x25      : chr  "Class0" "Class0" "Class2" "Class0" ...
#>  $ x26      : chr  "Class0" "Class0" "Class0" "Class3" ...
#>  $ x27      : chr  "Class0" "Class0" "Class2" "Class0" ...
#>  $ x28      : chr  "Class3" "Class0" "Class3" "Class0" ...
#>  $ x29      : chr  "Class0" "Class0" "Class2" "Class0" ...
#>  $ x30      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x31      : chr  "Class0" "Class0" "Class0" "Class0" ...
#>  $ x32      : chr  "Class1" "Class1" "Class2" "Class3" ...
#>  $ x33      : chr  "Class0" "Class0" "Class3" "Class0" ...
#>  $ x34      : int  55 8 26 40 45 41 18 57 22 30 ...
#>  $ target   : chr  "two" "one" "three" "one" ...

# 2. Develop and save model
set.seed(42)
p <- SupervisedModelDevelopmentParams$new()
p$df <- df
p$type <- "multiclass"
p$impute <- TRUE
p$grainCol <- "PatientID"
p$predictedCol <- "target"
p$debug <- FALSE
p$cores <- 1
# xgb_params must be a list with all of these things in it. 
# if you would like to tweak parameters, go for it! 
# Leave objective and eval_metric as they are.
p$xgb_params <- list("objective" = "multi:softprob",
                  "eval_metric" = "mlogloss",
                  "max_depth" = 6, # max depth of each learner
                  "eta" = 0.1, # learning rate
                  "silent" = 0, # verbose output when set to 1
                  "nthread" = 2) # number of processors to use

# Run model
boost <- XGBoostDevelopment$new(p)

#> Warning: Each of the following categorical variable levels occurs 3 times or fewer:
#> -  x6 : Class1
#> -  x27 : Class1
#> -  x33 : Class1
#> Consider grouping these together with other levels.

#> xgb_params for model training are: 
#> List of 7
#>  $ objective  : chr "multi:softprob"
#>  $ eval_metric: chr "mlogloss"
#>  $ max_depth  : num 6
#>  $ eta        : num 0.1
#>  $ silent     : num 0
#>  $ nthread    : num 2
#>  $ num_class  : int 6
#>  

boost$run()

#> Preparing data... 
#> Building model... 
#> Generating predictions... 
#> Generating confusion matrix... 
#> Confusion Matrix and Statistics
#> 
#>          True
#> Predicted one two five six three four
#>     one    22   0    0   0     0    0
#>     two     0  12    0   0     0    1
#>     five    0   0   10   0     1    0
#>     six     0   0    0   4     0    0
#>     three   0   0    0   0    13    0
#>     four    0   0    0   0     0    8
#> 
#> Overall Statistics
#>                                           
#>                Accuracy : 0.9718          
#>                  95% CI : (0.9019, 0.9966)
#>     No Information Rate : 0.3099          
#>     P-Value [Acc > NIR] : < 2.2e-16       
#>                                           
#>                   Kappa : 0.9647          
#>  Mcnemar's Test P-Value : NA              
#> 
#> Statistics by Class:
#> 
#>                      Class: one Class: two Class: five Class: six Class: three
#> Sensitivity              1.0000     1.0000      1.0000    1.00000       0.9286
#> Specificity              1.0000     0.9831      0.9836    1.00000       1.0000
#> Pos Pred Value           1.0000     0.9231      0.9091    1.00000       1.0000
#> Neg Pred Value           1.0000     1.0000      1.0000    1.00000       0.9828
#> Prevalence               0.3099     0.1690      0.1408    0.05634       0.1972
#> Detection Rate           0.3099     0.1690      0.1408    0.05634       0.1831
#> Detection Prevalence     0.3099     0.1831      0.1549    0.05634       0.1831
#> Balanced Accuracy        1.0000     0.9915      0.9918    1.00000       0.9643
#>                      Class: four
#> Sensitivity               0.8889
#> Specificity               1.0000
#> Pos Pred Value            1.0000
#> Neg Pred Value            0.9841
#> Prevalence                0.1268
#> Detection Rate            0.1127
#> Detection Prevalence      0.1127
#> Balanced Accuracy         0.9444

# Get output data 
outputDF <- boost$getPredictions()

#> Retrieving raw predictions... 

head(outputDF)

#>   PatientID           1           2           3           4           5
#> 1        11 0.002628169 0.003300496 0.983649611 0.002580024 0.002659348
#> 2        12 0.002686135 0.004936296 0.003321062 0.002636930 0.002718003
#> 3        20 0.978579640 0.007549843 0.004025320 0.002914262 0.003003862
#> 4        21 0.004226820 0.005051846 0.005920490 0.966682076 0.004276966
#> 5        24 0.002629580 0.004481775 0.003111624 0.002581409 0.974813104
#> 6        27 0.004178556 0.013649285 0.005227959 0.955643952 0.006535803
#>             6 predicted_label true_label
#> 1 0.005182362             one        one
#> 2 0.983701587             two        two
#> 3 0.003927014            five       five
#> 4 0.013841759             six        six
#> 5 0.012382531           three      three
#> 6 0.014764513             six        six

print(proc.time() - ptm)

#>    user  system elapsed 
#>   0.357   0.004   0.221