Train models without tuning for performance

Train models without tuning for performance

flash_models(d, outcome, model_class, models = c("rf", "knn"), n_folds = 5,
  hyperparameters, metric)

Arguments

d
A data frame

outcome
Name of the column to predict

model_class
"regression" or "classification". If not provided, this will be determined by the class of `outcome` with the determination displayed in a message.

models
Names of models to try, by default "rf" for random forest and "knn" for k-nearest neighbors. See supported_models for available models.

n_folds
How many folds to train the model on. Default = 5, minimum = 2. Whie flash_models doesn't use cross validation to tune hyperparameters, it trains n_folds models to evaluate performance out of fold.

hyperparameters
Optional list of hyperparameters to use. If missing, default values will be used. If provided, must be a named list of named lists where the outer list contains models and the inner lists contain hyperparameter values. See healthcareai:::get_hyperparameter_defaults() for a template.

metric
What metric to use to assess model performance? Options for regression: "RMSE" (root-mean-squared error, default), "MAE" (mean-absolute error), or "Rsquared." For classification: "ROC" (area under the receiver operating characteristic curve), or "PR" (area under the precision-recall curve).

Value

A model_list object

Details

This function has two major differences from tune_models:

It is much faster.

See also

Examples

# Prepare data
prepped_data <- prep_data(pima_diabetes, patient_id, outcome = diabetes)

#> Training new data prep recipe

# Simplest use. Get models quickly at default hyperparameter values
flash_models(prepped_data, diabetes)

#> Variable(s) ignored in prep_data won't be used to tune models: patient_id

#> diabetes looks categorical, so training classification algorithms.

#> Algorithms Trained: Random Forest, k-Nearest Neighbors
#> Target: diabetes
#> Class: Classification
#> Performance Metric: ROC
#> Number of Observations: 768
#> Number of Features: 12
#> Models Trained: 2018-03-30 14:25:28 
#> 
#> Models have not been tuned. Performance estimated via 5-fold cross validation at fixed hyperparameter values.
#> Best model: Random Forest
#> ROC = 0.84
#> User-selected hyperparameter values:
#>   mtry = 5
#>   splitrule = extratrees
#>   min.node.size = 10 

# Set non-default hyperparameter values by passing a list of lists to \code{hyperparameters}
models <-
  flash_models(d = prepped_data,
               outcome = diabetes,
               hyperparameters = list(
                 rf = list(
                   mtry = 3,
                   splitrule = "gini",
                   min.node.size = 1
                 ),
                 knn = list(
                   kmax = 3,
                   distance = 2,
                   kernel = "gaussian"
                 )
               )
  )

#> Variable(s) ignored in prep_data won't be used to tune models: patient_id

#> diabetes looks categorical, so training classification algorithms.

summary(models)

#> Models trained: 2018-03-30 14:25:30
#> 
#> Models have not been tuned. Performance estimated via 5-fold cross validation at fixed hyperparameter values.
#> Best algorithm: Random Forest with ROC = 0.84
#> 
#> Out-of-fold performance of all trained models:
#> 
#> $`Random Forest`
#>   mtry splitrule min.node.size       ROC  Sens      Spec      ROCSD     SensSD
#> 1    3      gini             1 0.8358218 0.868 0.5895178 0.03131845 0.04658326
#>       SpecSD
#> 1 0.06015438
#> 
#> $`k-Nearest Neighbors`
#>   kmax distance   kernel       ROC  Sens      Spec      ROCSD     SensSD
#> 1    3        2 gaussian 0.7835238 0.828 0.5934312 0.01653328 0.02863564
#>       SpecSD
#> 1 0.05642804
#> 

# Speed comparison of no tuning with flash_models vs. tuning with tune_models:

# NOT RUN {
  # ~40 seconds:
  system.time(
    tune_models(prepped_data, diabetes)
  )
  # ~6 seconds:
  system.time(
    flash_models(prepped_data, diabetes)
  )
# }