healthcareai: Streamlined Machine Learning

healthcare-ai provides a clean interface to create and compare multiple models on your data and then deploy the model that is most accurate. healthcareai also includes functions for data exploration, data cleaning, and model evaluation.

Details

Modeling is done in a three-step process: First, loading, understanding, and preparing the data. Second, tuning and selecting a best model. Third, making predictions, deploying the model, and monitoring over time.

Load data
selectData to pull data directly from a SQL database

Profile data
control_chart to plot data over time. pivot if you multiple rows for each observation you want a prediction on (e.g. multiple lab values for patients on separate rows), pivot can help reshape the data.

Prepare data for modeling
missingness to examine how much of each variable is missing. impute to impute missing values, or... prep_data to impute and perform additional data manipulation to get data ready for machine learning. split_train_test to split data into one data frame for developing a model and another for evaluating the model.

Tune machine learning models
machine_learn does all the work for you, from imputing missing values to finding optimal models via cross validation. If you want to keep the guts under the hood (apologies for the mixed metaphor!), this is the function for you! tune_models gives you more control over how models are trained.

Assess models
machine_learn and tune_models both return model_list objects. You can get high level information about your models by simply printing the model_list object (i.e. type the name of the variable holding the models). summary(model_list) provides even more information about model tuning. plot(model_list) plots model performance over hyperparameter space.

Make predictions and deploy the model
predict.model_list generates predictions writeData to push the predicted values into a SQL environment.

References

https://github.com/HealthCatalyst/healthcareai-r/

http://healthcare.ai