Machine learning made easy

Prepare data and train machine learning models.

machine_learn(d, ..., outcome, models, tune = TRUE, positive_class,
  n_folds = 5, tune_depth = 10, impute = TRUE)

Arguments

d
A data frame

...
Columns to be ignored in model training, e.g. ID columns, unquoted.

outcome
Name of the target column, i.e. what you want to predict. Unquoted. Must be named, i.e. you must specify outcome =

models
Models to be trained, k-nearest neighbors and random forest by default. See supported_models for details.

tune
If TRUE (default) models will be tuned via tune_models. If FALSE, models will be trained via flash_models which is substantially faster but produces less-predictively powerful models.

positive_class
For classification only, which outcome level is the "yes" case, i.e. should be associated with high probabilities? Defaults to "Y" or "yes" if present, otherwise is the first level of the outcome variable (first alphabetically if the training data outcome was not already a factor).

n_folds
How many folds to use to assess out-of-fold accuracy? Default = 5. Models are evaluated on out-of-fold predictions whether tune is TRUE or FALSE.

tune_depth
How many hyperparameter combinations to try? Defualt = 10. Ignored if tune is FALSE.

impute
Logical, if TRUE (default) missing values will be filled by hcai_impute

Value

A model_list object. You can call plot, summary, evaluate, or predict on a model_list.

Details

This is a high-level wrapper function. For finer control of data cleaning and preparation use prep_data or the functions it wraps. For finer control of model tuning use tune_models.

Examples

# Split the data into training and test sets, using just 100 rows for speed
d <- split_train_test(d = pima_diabetes[1:100, ],
                      outcome = diabetes,
                      percent_train = .9)

### Classification ###

# Clean and prep the training data, specifying that patient_id is an ID column,
# and tune algorithms over hyperparameter values to predict diabetes
diabetes_models <- machine_learn(d$train, patient_id, outcome = diabetes)

#> Training new data prep recipe

#> Variable(s) ignored in prep_data won't be used to tune models: patient_id

#> diabetes looks categorical, so training classification algorithms.

#> Training with cross validation: Random Forest

#> Training with cross validation: k-Nearest Neighbors

#> 
#> *** Models successfully trained. The model object contains the training data minus ignored ID columns. ***
#> *** If there was PHI in training data, normal PHI protocols apply to the model object. ***

# Inspect model specification and performance
diabetes_models

#> Algorithms Trained: Random Forest, k-Nearest Neighbors
#> Target: diabetes
#> Class: Classification
#> Performance Metric: AUROC
#> Number of Observations: 91
#> Number of Features: 12
#> Models Trained: 2018-04-20 13:20:13 
#> 
#> Models tuned via 5-fold cross validation over 10 combinations of hyperparameter values.
#> Best model: Random Forest
#> AUPR = 0.56, AUROC = 0.77
#> Optimal hyperparameter values:
#>   mtry = 1
#>   splitrule = extratrees
#>   min.node.size = 19 

# Make predictions (predicted probability of diabetes) on test data
predict(diabetes_models, d$test)

#> Prepping data based on provided recipe

#> "predicted_diabetes" predicted by Random Forest last trained: 2018-04-20 13:20:13
#> Performance in training: AUROC = 0.77

#> # A tibble: 9 x 11
#>   diabetes predicted_diabet… patient_id pregnancies plasma_glucose diastolic_bp
#> * <chr>                <dbl>      <int>       <int>          <int>        <int>
#> 1 Y                    0.359          7           3             78           50
#> 2 Y                    0.398         17           0            118           84
#> 3 N                    0.501         31           5            109           75
#> 4 N                    0.292         36           4            103           60
#> 5 N                    0.284         47           1            146           56
#> 6 N                    0.284         70           4            146           85
#> 7 Y                    0.397         73          13            126           90
#> 8 N                    0.422         78           5             95           72
#> 9 N                    0.279         83           7             83           78
#> # ... with 5 more variables: skinfold <int>, insulin <int>, weight_class <chr>,
#> #   pedigree <dbl>, age <int>

### Regression ###

# If the outcome variable is numeric, regression models will be trained
age_model <- machine_learn(d$train, patient_id, outcome = age)

#> Training new data prep recipe

#> Variable(s) ignored in prep_data won't be used to tune models: patient_id

#> age looks numeric, so training regression algorithms.

#> Training with cross validation: Random Forest

#> Training with cross validation: k-Nearest Neighbors

#> 
#> *** Models successfully trained. The model object contains the training data minus ignored ID columns. ***
#> *** If there was PHI in training data, normal PHI protocols apply to the model object. ***

# If new data isn't specifed, get predictions on training data
predict(age_model)

#> "predicted_age" predicted by Random Forest last trained: 2018-04-20 13:20:17
#> Performance in training: RMSE = 9.04

#> # A tibble: 91 x 16
#>      age predicted_age pregnancies plasma_glucose diastolic_bp skinfold insulin
#>  * <int>         <dbl>       <int>          <dbl>        <dbl>    <dbl>   <dbl>
#>  1    50          40.1           6           148.         72.0     35.0   169. 
#>  2    31          26.1           1            85.         66.0     29.0   169. 
#>  3    32          44.8           8           183.         64.0     29.0   169. 
#>  4    21          26.6           1            89.         66.0     23.0    94.0
#>  5    33          34.0           0           137.         40.0     35.0   168. 
#>  6    30          33.0           5           116.         74.0     29.0   169. 
#>  7    29          39.6          10           115.         72.4     29.0   169. 
#>  8    53          40.0           2           197.         70.0     45.0   543. 
#>  9    54          38.2           8           125.         96.0     29.0   169. 
#> 10    30          34.2           4           110.         92.0     29.0   169. 
#> # ... with 81 more rows, and 9 more variables: pedigree <dbl>,
#> #   weight_class_normal <dbl>, weight_class_obese <dbl>,
#> #   weight_class_overweight <dbl>, weight_class_missing <dbl>,
#> #   weight_class_other <dbl>, diabetes_Y <dbl>, diabetes_other <dbl>,
#> #   diabetes_missing <dbl>

### Faster model training without tuning hyperparameters ###

# Train models at set hyperparameter values by setting tune to FALSE. This is
# faster (especially on larger datasets), but produces models with less
# predictive accuracy.
machine_learn(d$train, patient_id, outcome = diabetes, tune = FALSE)

#> Training new data prep recipe

#> Variable(s) ignored in prep_data won't be used to tune models: patient_id

#> diabetes looks categorical, so training classification algorithms.

#> Training at fixed values: Random Forest

#> Training at fixed values: k-Nearest Neighbors

#> 
#> *** Models successfully trained. The model object contains the training data minus ignored ID columns. ***
#> *** If there was PHI in training data, normal PHI protocols apply to the model object. ***

#> Algorithms Trained: Random Forest, k-Nearest Neighbors
#> Target: diabetes
#> Class: Classification
#> Performance Metric: AUROC
#> Number of Observations: 91
#> Number of Features: 12
#> Models Trained: 2018-04-20 13:20:19 
#> 
#> Models have not been tuned. Performance estimated via 5-fold cross validation at fixed hyperparameter values.
#> Best model: Random Forest
#> AUPR = 0.57, AUROC = 0.78
#> User-selected hyperparameter values:
#>   mtry = 3
#>   splitrule = extratrees
#>   min.node.size = 1