Machine Learn

Prepare data and train machine learning models.

machine_learn(d, ..., outcome, models, tune = TRUE, n_folds = 5,
  tune_depth = 10, impute = TRUE)

Arguments

d
A data frame

...
Columns to be ignored in model training, e.g. ID columns, unquoted.

outcome
Name of the target column, i.e. what you want to predict. Unquoted. Must be named, i.e. you must specify outcome =

models
Models to be trained, k-nearest neighbors and random forest by default. See supported_models for details.

tune
If TRUE (default) models will be tuned via tune_models. If FALSE, models will be trained via flash_models which is substantially faster but produces less-predictively powerful models.

n_folds
How many folds to use to assess out-of-fold accuracy? Default = 5. Models are evaluated on out-of-fold predictions whether tune is TRUE or FALSE.

tune_depth
How many hyperparameter combinations to try? Defualt = 10. Ignored if tune is FALSE.

impute
Logical, if TRUE (default) missing values will be filled by hcai_impute

Value

model_list object ready to make predictions via predict.model_list

Details

This is a high-level wrapper function. For finer control of data cleaning and preparation use prep_data or the functions it wraps. For finer control of model tuning use tune_models.

Examples

# Split data into training and test sets using a subset of the data for speed
training_data <- pima_diabetes[1:50, ]
test_data <- pima_diabetes[51:60, ]

### Classification ###

# Clean and prep the data, tune algorithms over hyperparameter values to predict diabetes
diabetes_models <- machine_learn(training_data, outcome = diabetes)

#> Training new data prep recipe

#> diabetes looks categorical, so training classification algorithms.

#> Running cross validation for Random Forest

#> Running cross validation for k-Nearest Neighbors

# Make predictions (predicted probability of diabetes) on test data
predict(diabetes_models, test_data)

#> Prepping data based on provided recipe

#> "predicted_diabetes" predicted by Random Forest last trained: 2018-03-30 16:22:07
#> Performance in training: ROC = 0.75

#> # A tibble: 10 x 11
#>    diabetes predicted_diabe… patient_id pregnancies plasma_glucose diastolic_bp
#>  * <chr>               <dbl>      <int>       <int>          <int>        <int>
#>  1 N                   0.255         51           1            103           80
#>  2 N                   0.278         52           1            101           50
#>  3 N                   0.260         53           5             88           66
#>  4 Y                   0.663         54           8            176           90
#>  5 N                   0.716         55           7            150           66
#>  6 N                   0.267         56           1             73           50
#>  7 Y                   0.706         57           7            187           68
#>  8 N                   0.558         58           0            100           88
#>  9 N                   0.503         59           0            146           82
#> 10 N                   0.432         60           0            105           64
#> # ... with 5 more variables: skinfold <int>, insulin <int>, weight_class <chr>,
#> #   pedigree <dbl>, age <int>

### Regression ###

# Predict numeric outcomes simply by specifying the name of the outcome variable
age_model <- machine_learn(training_data, outcome = age)

#> Training new data prep recipe

#> age looks numeric, so training regression algorithms.

#> Running cross validation for Random Forest

#> Running cross validation for k-Nearest Neighbors

# If new data isn't specifed, get predictions on training data. Plot predictions
predict(age_model)

#> "predicted_age" predicted by Random Forest last trained: 2018-03-30 16:22:10
#> Performance in training: RMSE = 9.88

#> # A tibble: 50 x 17
#>      age predicted_age patient_id pregnancies plasma_glucose diastolic_bp
#>  * <int>         <dbl>      <int>       <int>          <int>        <dbl>
#>  1    50          44.1          1           6            148         72.0
#>  2    31          29.0          2           1             85         66.0
#>  3    32          36.4          3           8            183         64.0
#>  4    21          25.1          4           1             89         66.0
#>  5    33          33.3          5           0            137         40.0
#>  6    30          32.5          6           5            116         74.0
#>  7    26          27.8          7           3             78         50.0
#>  8    29          33.6          8          10            115         73.6
#>  9    53          47.8          9           2            197         70.0
#> 10    54          46.9         10           8            125         96.0
#> # ... with 40 more rows, and 11 more variables: skinfold <dbl>, insulin <dbl>,
#> #   pedigree <dbl>, weight_class_normal <dbl>, weight_class_obese <dbl>,
#> #   weight_class_overweight <dbl>, weight_class_hcai_missing <dbl>,
#> #   weight_class_other <dbl>, diabetes_Y <dbl>, diabetes_other <dbl>,
#> #   diabetes_hcai_missing <dbl>

### Faster model training without tuning hyperparameters ###

# Train models at set hyperparameter values by setting tune to FALSE.
# This is faster (especially on larger datasets), but produces models with less predictive accuracy.
machine_learn(training_data, outcome = diabetes, tune = FALSE)

#> Training new data prep recipe

#> diabetes looks categorical, so training classification algorithms.

#> Algorithms Trained: Random Forest, k-Nearest Neighbors
#> Target: diabetes
#> Class: Classification
#> Performance Metric: ROC
#> Number of Observations: 50
#> Number of Features: 13
#> Models Trained: 2018-03-30 16:22:11 
#> 
#> Models have not been tuned. Performance estimated via 5-fold cross validation at fixed hyperparameter values.
#> Best model: Random Forest
#> ROC = 0.67
#> User-selected hyperparameter values:
#>   mtry = 5
#>   splitrule = extratrees
#>   min.node.size = 10