Tune multiple machine learning models using cross validation to optimize performance

Tune multiple machine learning models using cross validation to optimize performance

tune_models(d, outcome, models, metric, positive_class, n_folds = 5,
  tune_depth = 10, tune_method = "random", hyperparameters, model_class)

Arguments

d
A data frame

outcome
Name of the column to predict

models
Names of models to try, by default "rf" for random forest and "knn" for k-nearest neighbors. See supported_models for available models.

metric
What metric to use to assess model performance? Options for regression: "RMSE" (root-mean-squared error, default), "MAE" (mean-absolute error), or "Rsquared." For classification: "ROC" (area under the receiver operating characteristic curve), or "PR" (area under the precision-recall curve).

positive_class
For classification only, which outcome level is the "yes" case, i.e. should be associated with high probabilities? Defaults to "Y" or "yes" if present, otherwise is the first level of the outcome variable (first alphabetically if the training data outcome was not already a factor).

n_folds
How many folds to use in cross-validation? Default = 5.

tune_depth
How many hyperparameter combinations to try? Defualt = 10.

tune_method
How to search hyperparameter space? Default = "random".

hyperparameters
Currently not supported.

model_class
"regression" or "classification". If not provided, this will be determined by the class of `outcome` with the determination displayed in a message.

Value

A model_list object

Details

Note that this function is training a lot of models (100 by default) and so can take a while to execute. In general a model is trained for each hyperparameter combination in each fold for each model, so run time is a function of length(models) x n_folds x tune_depth. At the default settings, a 1000 row, 10 column data frame should complete in about 30 seconds on a good laptop.

See also

prep_data, predict.model_list, supported_models

Examples

# NOT RUN {
### Takes ~20 seconds
# Prepare data for tuning
d <- prep_data(pima_diabetes, patient_id, outcome = diabetes)

# Tune random forest and k-nearest neighbors classification models
m <- tune_models(d, outcome = diabetes)

# Get some info about the tuned models
m

# Get more detailed info
summary(m)

# Plot performance over hyperparameter values for each algorithm
plot(m)

# Extract confusion matrix for random forest (the model with best-performing
# hyperparameter values is used)
caret::confusionMatrix(m$`Random Forest`, norm = "none")

# Compare performance of algorithms at best hyperparameter values
rs <- resamples(m)
dotplot(rs)
# }